Prepare full board

+++ PlayerInformations.h
@@ -6,7 +6,7 @@ class PlayerInformations
 public:
 	int id;
 	string name;
-	int money;
+	float money;
 	int ownedPlaces[40];
 	int position;
